filebuffer: implement
It was decided to create a new way to store data about the file and work with it, since at the moment the implementation of everything was made in the form of bicycles. Now a structure has been created that describes the basic ability to operate with old and new array slices of files and support for additional things.