storage: implement
Storage is an interface that allows you to create a logical structure to support new cloud services.
Upload is used for uploading files in the cloud storage. The name of the file that you upload must be in the form of ID is uuid.
Download is used for downloading files. Data about the file must be saved to the same structure using the Write or WriteFromReader method.